+++ lib/rex/io/datagram_abstraction.rb
@@ -1,6 +1,6 @@
 # -*- coding: binary -*-
 
-require 'socket'
+require 'rex/io/socket_abstraction'
 
 module Rex
 module IO
@@ -12,23 +12,13 @@ module IO
 #
 ###
 module DatagramAbstraction
-
+  include Rex::IO::SocketAbstraction
   #
   # Creates a streaming socket pair
   #
   def initialize_abstraction
-    self.lsock, self.rsock = Rex::Socket.udp_socket_pair()
+    self.lsock, self.rsock = Rex::Socket.udp_socket_pair
   end
-
-
-  # The left side of the stream (local)
-  attr_reader :lsock
-  # The right side of the stream (remote)
-  attr_reader :rsock
-
-protected
-  attr_writer :lsock
-  attr_writer :rsock
 
 end
