+++ lib/rex/post/meterpreter/extensions/stdapi/net/socket_subsystem/tcp_client_channel.rb
@@ -23,61 +23,6 @@ module SocketSubsystem
 ###
 class TcpClientChannel < Rex::Post::Meterpreter::Stream
 
-  class << self
-    def cls
-      return CHANNEL_CLASS_STREAM
-    end
-  end
-
-  module SocketInterface
-    def type?
-      'tcp'
-    end
-
-    def getsockname
-      return super if not channel
-      # Find the first host in our chain (our address)
-      hops = 0
-      csock = channel.client.sock
-      while(csock.respond_to?('channel'))
-        csock = csock.channel.client.sock
-        hops += 1
-      end
-      tmp,caddr,cport = csock.getsockname
-      tmp,raddr,rport = csock.getpeername
-      maddr,mport = [ channel.params.localhost, channel.params.localport ]
-      [ tmp, "#{caddr}#{(hops > 0) ? "-_#{hops}_" : ""}-#{raddr}", "#{mport}" ]
-    end
-
-    def getpeername
-      return super if not channel
-      tmp,caddr,cport = channel.client.sock.getpeername
-      maddr,mport = [ channel.params.peerhost, channel.params.peerport ]
-      [ tmp, "#{maddr}", "#{mport}" ]
-    end
-
-    attr_accessor :channel
-  end
-
-  #
-  # Simple mixin for lsock in order to help avoid a ruby interpreter issue with ::Socket.pair
-  # Instead of writing to the lsock, reading from the rsock and then writing to the channel,
-  # we use this mixin to directly write to the channel.
-  #
-  # Note: This does not work with OpenSSL as OpenSSL is implemented natively and requires a real
-  # socket to write to and we cant intercept the sockets syswrite at a native level.
-  #
-  # Note: The deadlock only seems to effect the Ruby build for cygwin.
-  #
-  module DirectChannelWrite
-
-    def syswrite( buf )
-      channel._write( buf )
-    end
-
-    attr_accessor :channel
-  end
-
   ##
   #
   # Factory
@@ -161,19 +106,6 @@ class TcpClientChannel < Rex::Post::Meterpreter::Stream
     return true
   end
 
-  #
-  # Wrap the _write() call in order to catch some common, but harmless Windows exceptions
-  #
-  def _write(*args)
-    begin
-      super(*args)
-    rescue ::Rex::Post::Meterpreter::RequestError => e
-      case e.code
-      when 10000 .. 10100
-        raise ::Rex::ConnectionError.new
-      end
-    end
-  end
 end
 
 end; end; end; end; end; end; end
